[FIX] web,mrp: move test to adequate module
