+++ addons/mrp/static/tests/mrp_tests.js
@@ -1,7 +1,6 @@
 odoo.define('mrp.tests', function (require) {
 "use strict";
 
-var field_registry = require('web.field_registry');
 var FormView = require('web.FormView');
 var testUtils = require("web.test_utils");
 
@@ -29,49 +28,6 @@ QUnit.module('mrp', {
         };
     },
 }, function () {
-
-    QUnit.test("pdf_viewer: upload rendering", function (assert) {
-        assert.expect(6);
-
-        testUtils.patch(field_registry.map.pdf_viewer, {
-            on_file_change: function (ev) {
-                ev.target = {files: [new Blob()]};
-                this._super.apply(this, arguments);
-            },
-            _getURI: function (fileURI) {
-                var res = this._super.apply(this, arguments);
-                assert.step('_getURI');
-                assert.ok(_.str.startsWith(fileURI, 'blob:'));
-                this.PDFViewerApplication = {
-                    open: function (URI) {
-                        assert.step('open');
-                        assert.ok(_.str.startsWith(URI, 'blob:'));
-                    },
-                };
-                return 'about:blank';
-            },
-        });
-
-        var form = createView({
-            View: FormView,
-            model: 'partner',
-            data: this.data,
-            arch:
-                '<form>' +
-                    '<field name="document" widget="pdf_viewer"/>' +
-                '</form>',
-        });
-
-        // first upload initialize iframe
-        form.$('input[type="file"]').trigger('change');
-        assert.verifySteps(['_getURI']);
-        // second upload call pdfjs method inside iframe
-        form.$('input[type="file"]').trigger('change');
-        assert.verifySteps(['_getURI', 'open']);
-
-        testUtils.unpatch(field_registry.map.pdf_viewer);
-        form.destroy();
-    });
 
     QUnit.test("bullet_state: basic rendering", function (assert) {
         assert.expect(2);
